Added support for a basic torchrun_hpc.py and associated stub file for launching distributed PyTorch jobs.  Currently it only uses the socket-based rendezvous protocol.
Fixed a bug in how launcher flags were inserted into the command line.

Added the ability for the CLI to prepend environment variables to the
system configuration.  This enabled the torchrun_hpc launcher to setup
the rendezvous protocol.

Added flag to the launcher launch function to provide the necessary
modifications for the torchrun_hpc CLI.