Added fastload option to reduce flux launch time.
Added preliminary implementation of using the mpi_rndv protocol for
initializing PyTorch distributed.

Added the mpi_rndv library as a dependency for the HPC launcher when
using MPI.  Disabled a recommended OFI policy flag that is causing
crashes on mi300a nodes, due to the `Launcher/PMI did not supply NIC
data`.

Added a flag to specify the rendezvous protocol.  Added some unit
tests.

Added a flag to the enable a selection of which rendezvous protocol to
use.  Added logging for the torchrun_hpc_stub file.  Added test to
check that the correct rendezvous policy was launched.

Fixed a bug where ElCap systems added the fastload option for slurm
scheduler wrapper scripts.  Changed the torchrun-hpc tests to cleanup
their working directories when successful.

Updated the selection of the torchrun-hpc rendezvous protocol so that
it is done in the launch shell script, not within a python subprocess
within the torchrun_hpc_stub.py file.  This prevents a nesting problem
on LSF systems where a jsrun is called from the python subprocess
within a jsrun.

Added method for capturing hostlist for LSF and SLURM

Fix the test for a running flux scheduler.

Fixed the LSF launcher so that it will properly call a bsub -Is when
on the head node and jsrun when in an allocation for interactive jobs.

Added tests for LSF and properly captured stderr from bsub -Is

Added a guard for slurm schedulers on LSF systems (Sierra and Lassen)
so that the -u flag isn't passed to srun because it is a proxy for
lrun, which doesn't support the -u unbuffered flag.

Check for Sierra-class systems.

Fixed LSF Sierra launcher arguments.

Fixed LSF launcher class to append launcher flags to parallel run command.

Cleaned up code with review feedback.