Improved autodetect (#26)
* Added the capability to autodetect number of CPUs and GPUs, GPU
memory, GPU architecture, NUMA layout, and scheduler.  When possible
the GenericSystem will take advantage of this.  Removed redundant
field in the system parameters class.

Changed the behavior of the torchrun-hpc launcher so that when a
single rank MPI rendezvous job is scheduled, it will always setup
torch distributed.  If a TCP rendezvous is used for a single rank job,
initialization is deferred or elided.

* Updated both the NVIDIA and NUMA autodetect to use native NVIDIA and
local code respectively.

* Improved how the autodetect works for AMD and NVIDIA systems so that
architecture specification is more robust.  Changed the data field for
GPU memory to a float.

* Apply suggestions from code review

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>

* Removed unused field.

* Apply suggestions from code review

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>

* Cleaned up the code.

* Ran black.

---------

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>