Add CLI option for system parameters and support for Matrix system (#27)
* Added a system definition for Matrix and Vector.

* Fixed memory size

* Added command line argument to express the desired number of GPUs per
process (task).  For most AI codes, this should be 1, which is the
default, but it can now be set.

Updated the FLUX and SLURM schedulers to set this field when
necessary.  This also addresses an issue when running on compute
resources that can be shared and are not exclusive.

* Ran black.

* Added support for specifying a set of command line arguments for the
system parameters.  These will overwrite and known or autodetected
system parameters.