Bugfix rccl libfabric (#33)
* Update the scheduler so that it always uses a full file path for the
command provided.

* Updated networking flags to fix interaction between RCCL and libfabric
systems.  Advised and audited by HPE and AMD.

* Updates copyright statements.

* Added a CLI argument that allows the user to specific environmental
optimizations for a jobs dominant communication protocol, e.g. MPI vs
*CCL (RCCL, NCCL).  Separated certain El Cap environmental
optimizations such that they are appropriately applied for the proper
communication protocol.  Improved the ability to add comments for
environmental variables and add comments to launch files.

* Fixed bug in local scheduler.

* Bugfix.

* Always export HPC-Launcher PyTorch distributed environment variables
to allow downstream packages to see them.

* Set default dummy values for MASTER_ADDR and MASTER_PORT if the MPI
rendezvous protocol is used.

* Ensure that the local rank environment variable is also set.

* Apply suggestions from code review

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>

* Disable setting the NCCL_SOCKET_IFNAME and fix a typo in it's name.

* Set the default torchrun-hpc rendezvous protocol to always default to
TCP.  Add a end to end test for PyTorch with the Huggingface
Accelerate library.

* Code beautification

* Remove unnecessary OMP env variables.

* Fixed how GPUs are allocated using LOCAL_RANK env variable.  Check the
actual local rank against the number of available devices and do
round-robin modulo scheduling.  This works both with mpibind and
without mpibind.

Fixed a bug in the El Cap system where an optional flux flag was not
picked up due to a space between the -o flag and the option.

---------

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>