Added markdown of the CLI help messages. (#51)
* Added markdown of the CLI help messages.

* Fix some bugs in the examples.

* Fixed a bug where the number of GPUs required was returned as a float
rather than an int.

* Fixed how banks and accounts are specified.

* Fixed a bug in how the check for None vs an empty string in the
configuration of the launch directory argument.

Updated some of the test configuration examples.

* Improved how the system architecture (especially if overridden) is reported.

* Added guards for setting out or error log files for ephemeral jobs.

* Added function to scheduler class to get the environment variable for
each rank's ID.  Updated the launch script so that it gets the RANK
environment variable so that it can write out the hostlist if
necessary.  Improved the guards for ephemeral job CLI flags.

* Removed debugging code

* Cleaned up and improved integration with the torchrun-hpc CLI argument
to set the max memory size and the CLI parameter list.  Fixed a bug in
how the system parameters are mutated from the CLI.

* Updated env variable.

* Added a default argument for the max gpu mem.

* Fixed how slurm runs check for the root node in a torch run.

* Finished cleaning up the torchrun-hpc CLI examples.

* Fixed tests to use a launch directory.

* Minor cleanup